requirements.txt
# EasyPay Payment Gateway - Core Dependencies (Updated December 2024)

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Redis
redis==5.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# HTTP Client
httpx==0.28.1
aiohttp==3.10.11

# Data Validation
pydantic==2.10.3
pydantic-settings==2.7.0

# Environment Management
python-dotenv==1.0.1

# Logging
structlog==24.4.0

# Monitoring
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Development Tools
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Payment Processing
authorizenet==1.1.6

# Utilities
python-dateutil==2.9.0
pytz==2024.2
psutil==6.1.0
python-json-logger==3.2.0
